refactor(mcp): extract the generation tool catalog out of mcpGenerationTools
mcpGenerationTools.ts 826 -> 666 and leaves the filesize allowlist. The declarative catalog
moves to mcpGenerationToolCatalog.ts: MCP_GENERATION_TOOL_CATALOG plus the two helpers only it
uses, `gstr` and OPERATION_PLAN_SHARED_FIELDS. The catalog is pure contract — tool names,
descriptions, input schemas and the build functions that fold MCP params into method calls. It
changes when the exposed tool surface changes, not when the handler implementation does. The
handler, the operation store and dispatch stay behind.

GENERATION_RECONCILE_OUTCOMES deliberately did NOT move, even though the catalog is its only
in-file consumer. It is a registered semantic-vocabulary owner with a baseline site, and it has
two external importers (modelToolSurfaceManifest.ts, generationTransportAdapters.ts). Moving it
would relocate a vocabulary site, which is a separate decision and should not ride along inside
a structural extraction. The new module imports it instead.

The three importers now point at the owner rather than going through mcpGenerationTools.

Verified: npx tsc -p electron/tsconfig.json EXIT=0; mcpGenerationTools + projectSessionDispatch
+ mcpGenerationDispatcher tests 84/84; pnpm run lint:ci EXIT=0;
node scripts/check-file-sizes.mjs EXIT=0; pnpm run check:vocabularies EXIT=0 (the vocabulary
owner stayed put, so no site moved).
