+++ electron/capabilityCore/projectSessionDispatch.test.ts
@@ -8,7 +8,7 @@ import { dispatch } from './dispatcher'
 import { isSemanticGenerationRoute } from './generationDispatcher'
 import type { McpConnectionContext } from './mcpConnectionContext'
 import { createMcpGenerationPolicy } from './mcpGenerationPolicy'
-import { MCP_GENERATION_TOOL_CATALOG } from './mcpGenerationTools'
+import { MCP_GENERATION_TOOL_CATALOG } from './mcpGenerationToolCatalog'
 import { createMcpProtocol, type McpTransport } from './mcpProtocol'
 import { MCP_TOOL_RESOLVER } from './mcpToolCatalog'
 import { createProjectLeaseAuthority } from './projectLease'
